requirements.txt
Flask==2.3.3
flask-cors==4.0.0
PyPDF2==3.0.1
google-generativeai==0.8.3
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==2.3.7
reportlab==4.2.2
